+++ src/me/corriekay/pokegoutil/GUI/controller/PokemonTableController.java
@@ -14,8 +14,8 @@ import javafx.scene.input.MouseButton;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.GridPane;
 import me.corriekay.pokegoutil.DATA.managers.PokemonBagManager;
-import me.corriekay.pokegoutil.DATA.models.Operations.Operation;
 import me.corriekay.pokegoutil.DATA.models.PokemonModel;
+import me.corriekay.pokegoutil.DATA.models.operations.Operation;
 import me.corriekay.pokegoutil.GUI.enums.ColumnID;
 import me.corriekay.pokegoutil.GUI.enums.OperationID;
 import me.corriekay.pokegoutil.utils.ConfigKey;
